+++ meta_package_manager/managers/apm.py
@@ -49,7 +49,7 @@ class APM(PackageManager):
 
     requirement = "1.0.0"
 
-    version_regex = r"apm\s+(?P<version>\S+)"
+    version_regexes = (r"apm\s+(?P<version>\S+)",)
     """
     .. code-block:: shell-session