+++ meta_package_manager/managers/scoop.py
@@ -37,7 +37,7 @@ class Scoop(PackageManager):
 
     requirement = "0.2.4"
 
-    version_regex = r"^v(?P<version>\S+)\s.+"
+    version_regexes = (r"^v(?P<version>\S+)\s.+",)
     """Search version at the start of a line.
 
     .. code-block:: pwsh-session